requirements.txt
openai
python-dotenv
streamlit
pandas
numpy
matplotlib
networkx
pypdf2
python-docx